main/ar_tryon/management/utils/ar_validation.py
import cv2
import numpy as np
from PIL import Image
import base64

class ARValidation:
    """Validation utilities for AR try-on"""
    
    @staticmethod
    def validate_pose(image_data):
        """Validate if user pose is suitable for AR try-on"""
        try:
            # Decode base64 image
            image_bytes = base64.b64decode(image_data.split(',')[1])
            image = Image.open(io.BytesIO(image_bytes))
            
            # Convert to OpenCV format
            cv_image = cv2.cvtColor(np.array(image), cv2.COLOR_RGB2BGR)
            
            # Simple pose validation (you'd integrate with a pose estimation model)
            height, width = cv_image.shape[:2]
            
            # Check if image is clear and person is visible
            gray = cv2.cvtColor(cv_image, cv2.COLOR_BGR2GRAY)
            
            # Check contrast and brightness
            contrast = gray.std()
            brightness = gray.mean()
            
            validation_result = {
                'is_valid': True,
                'confidence': 0.8,
                'issues': []
            }
            
            if contrast < 30:
                validation_result['issues'].append('Image too blurry')
                validation_result['confidence'] -= 0.2
            
            if brightness < 50:
                validation_result['issues'].append('Image too dark')
                validation_result['confidence'] -= 0.2
            
            if brightness > 200:
                validation_result['issues'].append('Image too bright')
                validation_result['confidence'] -= 0.2
            
            validation_result['is_valid'] = validation_result['confidence'] > 0.5
            
            return validation_result
            
        except Exception as e:
            return {
                'is_valid': False,
                'confidence': 0,
                'issues': [f'Image processing error: {str(e)}']
            }
    
    @staticmethod
    def validate_lighting(image_data):
        """Validate lighting conditions for AR"""
        # Similar to pose validation but focused on lighting
        pass
    
    @staticmethod
    def validate_background(image_data):
        """Validate background for AR overlay"""
        # Check if background is suitable for AR
        pass